Adjusting groot root func

+++ root/Shell/chroot.sh
@@ -1,31 +1,34 @@
 function achroot {
-  mount /dev/sdb3 /mnt
-  mount -t proc /proc /mnt/proc
-  mount --rbind /dev /mnt/dev
-  mount --rbind /sys /mnt/sys
-  mount --rbind /run /mnt/run
-  cp /etc/resolv.conf /mnt/et
-  chroot /mnt /bin/bash
+    mount /dev/sdb3 /mnt
+    mount -t proc /proc /mnt/proc
+    mount --rbind /dev /mnt/dev
+    mount --rbind /sys /mnt/sys
+    mount --rbind /run /mnt/run
+    cp /etc/resolv.conf /mnt/et
+    chroot /mnt /bin/bash
 }
 
 function mchroot {
-  mount /dev/sdb1 /mnt
-  mount -t proc /proc /mnt/proc
-  mount --rbind /dev /mnt/dev
-  mount --rbind /sys /mnt/sys
-  mount --rbind /run /mnt/run
-  cp /etc/resolv.conf /mnt/et
-  chroot /mnt /bin/bash
+    mount /dev/sdb1 /mnt
+    mount -t proc /proc /mnt/proc
+    mount --rbind /dev /mnt/dev
+    mount --rbind /sys /mnt/sys
+    mount --rbind /run /mnt/run
+    cp /etc/resolv.conf /mnt/et
+    chroot /mnt /bin/bash
 }
 
 function groot {
-    mount -t proc /proc /gentoo/proc
-    mount --rbind /dev /gentoo/dev
-    mount --make-rslave /gentoo/dev
-    mount --rbind /sys /gentoo/sys
-    mount --make-rslave /gentoo/sys
-    rm /gentoo/etc/resolv.conf
-    cp /etc/resolv.conf /gentoo/etc
-    chroot /gentoo /usr/local/bin/md
-    chroot /gentoo /bin/zsh
+    if ! [[ -f /gentoo/proc/config.gz ]]; then
+         mount -t proc /proc /gentoo/proc
+         mount --rbind /dev /gentoo/dev
+         mount --make-rslave /gentoo/dev
+         mount --rbind /sys /gentoo/sys
+         mount --make-rslave /gentoo/sys
+         rm /gentoo/etc/resolv.conf
+         cp /etc/resolv.conf /gentoo/etc
+         chroot /gentoo /usr/local/bin/md
+    fi
+
+    chroot /gentoo /usr/local/bin/su-fusion809
 }
